Nearly complete Input code to segment a video copy it to the correct directory and then process all the segments

+++ python_video_processing.py
@@ -28,7 +28,7 @@ def main():
         print("Can't determine video length.")
         raise SystemExit
 
-    split_count = math.ceil(video_length / split_length)
+    split_count = int(math.ceil(video_length / split_length))
 
     if split_count == 1:
         print("Video length is less than the target split length.")
